Fix can`t update the properties of a model that has an indirect reference
Before the update, upsertModel will recalculate all properties of the model with updated data. At this time, if the indirectly referenced value has a value, the indirectly referenced reference will be undefined(even if the indirectly referenced value does not exist in data), resulting in the use of undefined to update the indirectly referenced value, throw an exception.